+++ CHANGELOG.md
@@ -5,7 +5,23 @@ All notable changes to the "EscapeNonAsciiToMD" extension will be documented in 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - Unreleased
+## [0.2.0] - 
+
+### Added
+- Enhanced visual feedback for lines containing non-ASCII characters.
+  - Added a red circle SVG icon in the gutter (numbers column) of lines with non-ASCII characters.
+- Introduced `highlightDecorationType` to apply a visual style (orange background with bold black text) for detected non-ASCII characters.
+- New command `Detect Invalid Chars` that highlights non-ASCII characters in the active editor (ID: `extension.detectinvalidchars`).
+- Added a `detectInvalidChars` function that identifies and highlights non-ASCII characters in the provided text editor.
+
+### Changed
+- Updated the SVG data encoding for gutter icons to be more efficient with base64 encoding.
+- Improved error handling for the default character mapping loading process.
+
+### Fixed
+- Resolved the issue with highlight text color where non-ASCII characters were not displayed with white text on a red background.
+
+## [0.1.0] - 
 
 ### Added
 - Initial release of the "EscapeNonAsciiToMD" extension.
@@ -34,4 +50,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 
 (Any deprecated features will be listed here in future updates)
-